requirements.txt
Django==4.2
drf-yasg==1.21.7
psycopg2-binary==2.9.7
python-dotenv==1.0.1
djangorestframework==3.14.0

